requirements.txt
streamlit>=1.27.0
requests>=2.31.0
pyautogen>=0.2.0
google-generativeai>=0.3.0
pandas>=2.0.0
matplotlib>=3.7.0